Don't include empty query parameter in URLs
The spreadsheets API endpoint returns an HTTP 400 error with a
message of "Invalid query parameter value for sq." when the
``sq`` query parameter is empty.  The default case for Tabletop
is an empty query.  Only add the ``sq`` parameter to the generated
URL if ``this.query`` is not empty.

Addresses https://github.com/jsoma/tabletop/issues/45