Fix to allow use with new version of Google Sheets
Now that pretty much everyone had moved over, or been moved over to the new Google Sheets, the URI reference given when a sheet is published has changed. This seems to work fine for me on sheets using the new version.
Also there was a double // in the original URL, and a pointless &sq= which broke things.
Ruby newbie, so apologies if code is rusty.